+++ cfg/cfg.go
@@ -27,7 +27,7 @@
 //       version: ^1.2.0
 //       repo:    git@github.com:Masterminds/vcs
 //       vcs:     git
-//     - package: github.com/codegangsta/cli
+//     - package: github.com/urfave/cli
 //     - package: github.com/Masterminds/semver
 //       version: ^1.0.0
 //